Renamed rule to action. Signed-off-by: Sebastian Prehn <sebastian.prehn@gmx.de> (github: sprehn)
