+++ addons/action/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>rule</name>
+	<name>action</name>
 	<comment></comment>
 	<projects>
 	</projects>
