+++ addons/action/org.openhab.action.lgwebos/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.openhab.rule.lgwebos</name>
+	<name>org.openhab.action.lgwebos</name>
 	<comment></comment>
 	<projects>
 	</projects>
