+++ addons/action/org.openhab.action.lgwebos/META-INF/MANIFEST.MF
@@ -1,13 +1,13 @@
 Manifest-Version: 1.0
 Bundle-ClassPath: .
 Bundle-ManifestVersion: 2
-Bundle-Name: LG webOS Rule
+Bundle-Name: LG webOS Action
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-SymbolicName: org.openhab.rule.lgwebos;singleton:=true
+Bundle-SymbolicName: org.openhab.action.lgwebos;singleton:=true
 Bundle-Vendor: openHAB
 Bundle-Version: 2.2.0.qualifier
 Export-Package: 
- org.openhab.rule.lgwebos.handler
+ org.openhab.action.lgwebos.handler
 Import-Package: 
  org.eclipse.jdt.annotation;resolution:=optional,
  org.eclipse.smarthome.automation,
@@ -24,7 +24,7 @@ Import-Package:
  org.eclipse.smarthome.io.console.extensions,
  org.eclipse.smarthome.model.item,
  org.openhab.binding.lgwebos,
- org.openhab.rule.lgwebos.handler,
+ org.openhab.action.lgwebos.handler,
  org.osgi.framework,
  org.osgi.service.component,
  org.osgi.service.event,
