+++ addons/action/org.openhab.action.lgwebos/about.html
@@ -31,9 +31,6 @@ did not receive this Content directly from the openHAB community, the following 
 for informational purposes only, and you should look to the Redistributor's license for 
 terms and conditions of use.</p>
      
-<ul>
-    <!--li><a href="https://github.com/sprehn/Connect-SDK-Java-Core/releases">Connect SDK Java Core</a> under <a href="http://www.apache.org/licenses/LICENSE-2.0.txt">Apache 2.0 license</a></li-->
-</ul>
-   
+<p>This bundle does not include any third party libraries.</p>   
 </body>
 </html>
