+++ addons/action/org.openhab.action.lgwebos/build.properties
