+++ addons/action/org.openhab.action.lgwebos/pom.xml
@@ -4,16 +4,16 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>org.openhab.rule</groupId>
+    <groupId>org.openhab.action</groupId>
     <artifactId>pom</artifactId>
     <version>2.2.0-SNAPSHOT</version>
   </parent>
 
-  <groupId>org.openhab.rule</groupId>
-  <artifactId>org.openhab.rule.lgwebos</artifactId>
+  <groupId>org.openhab.action</groupId>
+  <artifactId>org.openhab.action.lgwebos</artifactId>
   <version>2.2.0-SNAPSHOT</version>
   
-  <name>LG webOS Rule</name>
+  <name>LG webOS Action</name>
   <packaging>eclipse-plugin</packaging>    
 
 </project>
