+++ addons/action/org.openhab.action.lgwebos/src/main/java/org/openhab/action/lgwebos/handler/ButtonActionHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.rule.lgwebos.handler;
+package org.openhab.action.lgwebos.handler;
 
 import java.util.Map;
 
@@ -14,21 +14,22 @@ import org.eclipse.smarthome.automation.Action;
 import org.eclipse.smarthome.automation.handler.ActionHandler;
 import org.eclipse.smarthome.automation.handler.BaseModuleHandler;
 import org.openhab.binding.lgwebos.LGWebOS;
+import org.openhab.binding.lgwebos.LGWebOS.Button;
 
 /**
- * This action handler allows to open a URL.
+ * This action handler allows to send a key press.
  *
  * @author Sebastian Prehn - initial contribution
  *
  */
-public class BrowserActionHandler extends BaseModuleHandler<Action> implements ActionHandler {
-    public static final String TYPE_ID = "lgwebos.BrowserAction";
+public class ButtonActionHandler extends BaseModuleHandler<Action> implements ActionHandler {
+    public static final String TYPE_ID = "lgwebos.ButtonAction";
     public static final String PARAM_THING_ID = "thingId";
-    public static final String PARAM_URL = "url";
+    public static final String PARAM_BUTTON = "button";
 
     private LGWebOS api;
 
-    public BrowserActionHandler(Action module, LGWebOS api) {
+    public ButtonActionHandler(Action module, LGWebOS api) {
         super(module);
         this.api = api;
     }
@@ -36,8 +37,9 @@ public class BrowserActionHandler extends BaseModuleHandler<Action> implements A
     @Override
     public Map<String, Object> execute(Map<String, Object> context) {
         String thingId = module.getConfiguration().get(PARAM_THING_ID).toString();
-        String url = module.getConfiguration().get(PARAM_URL).toString();
-        api.launchBrowser(thingId, url);
+        Button button = Button.valueOf(module.getConfiguration().get(PARAM_BUTTON).toString());
+        api.sendButton(thingId, button);
         return null;
     }
+
 }
