+++ addons/action/org.openhab.action.lgwebos/src/main/java/org/openhab/action/lgwebos/handler/TextActionHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.rule.lgwebos.handler;
+package org.openhab.action.lgwebos.handler;
 
 import java.util.Map;
 
@@ -14,22 +14,21 @@ import org.eclipse.smarthome.automation.Action;
 import org.eclipse.smarthome.automation.handler.ActionHandler;
 import org.eclipse.smarthome.automation.handler.BaseModuleHandler;
 import org.openhab.binding.lgwebos.LGWebOS;
-import org.openhab.binding.lgwebos.LGWebOS.Button;
 
 /**
- * This action handler allows to send a key press.
+ * This action handler allows to send text input.
  *
  * @author Sebastian Prehn - initial contribution
  *
  */
-public class ButtonActionHandler extends BaseModuleHandler<Action> implements ActionHandler {
-    public static final String TYPE_ID = "lgwebos.ButtonAction";
+public class TextActionHandler extends BaseModuleHandler<Action> implements ActionHandler {
+    public static final String TYPE_ID = "lgwebos.TextAction";
     public static final String PARAM_THING_ID = "thingId";
-    public static final String PARAM_BUTTON = "button";
+    public static final String PARAM_TEXT = "text";
 
     private LGWebOS api;
 
-    public ButtonActionHandler(Action module, LGWebOS api) {
+    public TextActionHandler(Action module, LGWebOS api) {
         super(module);
         this.api = api;
     }
@@ -37,8 +36,8 @@ public class ButtonActionHandler extends BaseModuleHandler<Action> implements Ac
     @Override
     public Map<String, Object> execute(Map<String, Object> context) {
         String thingId = module.getConfiguration().get(PARAM_THING_ID).toString();
-        Button button = Button.valueOf(module.getConfiguration().get(PARAM_BUTTON).toString());
-        api.sendButton(thingId, button);
+        String text = module.getConfiguration().get(PARAM_TEXT).toString();
+        api.sendText(thingId, text);
         return null;
     }
